main_app/requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
psycopg2-binary
pyjwt
bcrypt

